+++ mithril-signer/src/runtime/mod.rs
@@ -1,9 +1,7 @@
 mod error;
 mod runner;
-mod signer_services;
 mod state_machine;
 
 pub use error::*;
 pub use runner::*;
-pub use signer_services::*;
 pub use state_machine::*;
